octofit-tracker/backend/requirements.txt
Django==3.1.14
djangorestframework==3.14.0
django-allauth==0.51.0
django-cors-headers==3.10.0
dj-rest-auth
djongo==1.3.6
pymongo==3.12
sqlparse==0.2.4